Embrace the lack of streaming iterators.
I don't see any relief for streaming iterators on the horizon. It seems
we are doomed until we get higher-kinded polymorphism.

In that case, we might as well make `next_field` a little more ergonomic.

Moreover, an extremely unsafe `byte_fields` iterator is provided. You
should only use it when absolutely necessary (for example, to process
CSV rows as iterators). WARNING: This is so ridiculously unsafe that it
is hidden in the public API. For example, if you call `collect` on the
iterator, then you will have wildly violated memory safety.