Tweak quotes so that it's possible to specify that quotes should never be used.
This is more useful in the encoder, where encoding a field that requires
quotes will report an error (only when `QuoteStyle::Never` is used).

On the parsing side, this could possibly lead to more efficient parsers.
(But I'm skeptical of anything but modest gains.)

As a result, the `quote` method has changed to accept an `Option`, so
this is a breaking change.

[breaking-change]