Rustup.
This required a couple breaking changes.

Firstly, we've moved to using `rustc_serialize` instead of `serialize`,
since the latter has been deprecated. This requires you to change:

    extern crate serialize;

to

    extern crate "rustc-serialize" as rustc_serialize;

And all instances of `deriving(Decodable)` and `deriving(Encodable)`
to `deriving(RustcDecodable)` and `deriving(RustcEncodable)`,
respectively.

The second breaking change is that several public methods used to
accept `S: StrAllocating`, which allowed you to pass a `String` or
a `&str`. Now that `StrAllocating` has been removed, this has been
switched to `Sized? S: ToOwned<String>`. This requires that `S` is
passed by reference. So change code like this:

    let s: String = String::from_str("hi");
    let mut rdr = csv::Reader::from_string(s);

to this

    let mut rdr = csv::Reader::from_string(&s);

[breaking-change]