Move to new std::io.
This is a large breaking change, but most of the fixes should be cosmetic.

First and foremost, the `IntoVector` and `StrAllocating` traits have been
removed. In all cases, they can be replaced with `IntoCow<'a, [u8]>` or
`IntoCow<'a, str>` bounds.

Secondly, `CsvResult` was renamed to `Result`, in keeping with the
conventions established in the standard library.

Thirdly, the `from_file` constructors no longer take a `Path`. They
now has an `AsPath + ?Sized` bound (which is the same used in the
standard library).

Fourthly, the `seek` method on `csv::Reader` now takes only a starting
offset. The seek style is removed (because only `io::SeekFrom::Start`
is supported).

Fifthly, migrating to `std::io` changes many bounds (e.g., `io::Read`
instead of `io::Reader`) and some types (e.g., `io::Curosr<Vec<u8>>`
instead of `io::MemReader`).

Finally, note that the streaming example has been removed because there
are no `ChanReader` or `ChanWriter` types defined for the new `std::io`.

[breaking-change]