csv/error: box the error.
The error type appears to have grown quite large and can result in lots of
byte copying. It's not clear exactly how much it costs, but boxing the
error results in less work.

Since we keep the box a secret, this adds a new ErrorKind that can be
extracted from a Error if one wants to do case analysis.

Since this rejiggers the Error type from a public enum to a private struct,
this is a breaking change.