csv-core: fix comment handling
This commit fixes a bug in how comments were interpreted. In particular,
this causes the behavior to match the documentation, which is correct.
Comments should only be allowed at the beginning of a record, but the
parser permitted them to also appear at the beginning of a field. This
corrects the implementation by simply moving the transition from
StartField to StartRecord, and then adds a couple regression tests.

Fixes #137