Implement serialize_header for serializing only the header row
Called "the simplest solution" in #161.

I concur the API is a bit weird, but the catches are documented and
a small example was added . There are probably more catches I don't know
about but what I do know I've documented.

Also, some tests were added to show some of the weirdness.

Closes #161.