Implement issue #167
This PR adds the option to specify a function to mean 'empty field' or
`None`, which currently is always assumed to be `b""`. It stores a
function on the heap, and then calls it with the raw contents of the
field to check for `None`ness.

Issues still to address

 - The name for the setting (I don't like the current name).
 - Testing
 - More documentation.

This commit serves as a proof of concept and a starting point.