api: automatically escape fields that contain the comment character
Previously, if data is written with QuoteStyle::Necessary, and the
first field of a row happens to contain a comment character, the row
will be ignored as a comment when later reading it back in.

This change adds a `comment` property to Writer, and automatically
quotes fields that have the provided comment character in them, so they
round-trip correctly.

Closes #283