error: drop 'source' impl
We can either write an underlying error's display string in our
own Display impl OR we can return it as a source. We can't do
both, otherwise we wind up with repeating error messages when using
today's error helper libraries (like anyhow).

Since the error types are either exposed or already offer a way
to get the underlying error, we just drop the 'source' impl.