Add IntoInnerError::into_error()
Currently, for code that doesn't attempt to recover from errors but just
reports the error to the caller, it's forced to return the entire
`IntoInnerError`, which is very big compared to `io::Error` which is
where the actual error is.

This commit adds `IntoInnerError::into_error()` method that consumes
`IntoInnerError` and returns contained `io::Error`. This mirrors the API
of `std::io::IntoInnerError` in Rust stdlib, which provides such method.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>