perf: serialize 128-bit integers via `itoa`
This optimizes the `serialize_i128` and `serialize_u128`
implementations by having them use `itoa`, as all of the other
integer implementations do. This avoids going through the `Display`
implementation of `u128`/`i128` and an intermediate allocation, which
is necessary because of how `collect_str` is currently implemented.

The current implementation seems to be a remnant of the pre-1.0 `itoa`
days, which didn't support `u128` and `i128`.

PR #401